+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#2323](https://github.com/open-telemetry/opentelemetry-python/pull/2323))
 - Complete metric exporter format and update OTLP exporter
   ([#2364](https://github.com/open-telemetry/opentelemetry-python/pull/2364))
-- [api] Add `_NoOpTracer` and `_NoOpTracerProvider`. Marking `_DefaultTracer` and `_DefaultTracerProvider` as deprecated.
+- [api] Add `NoOpTracer` and `NoOpTracerProvider`. Marking `_DefaultTracer` and `_DefaultTracerProvider` as deprecated.
   ([#2363](https://github.com/open-telemetry/opentelemetry-python/pull/2363))
 
 ## [1.8.0-0.27b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.8.0-0.27b0) - 2021-12-17